For Discovey agent client initialization use port value as set in agent param.
Closes-Bug: #1580109

+++ src/vnsw/agent/init/agent_param.cc
@@ -297,7 +297,7 @@ void AgentParam::ParseVirtualHost() {
 
 void AgentParam::ParseDiscovery() {
     GetValueFromTree<string>(dss_server_, "DISCOVERY.server");
-    GetValueFromTree<uint16_t>(dss_port_, "DISCOVERY.port");
+    GetValueFromTree<uint32_t>(dss_port_, "DISCOVERY.port");
     if (!GetValueFromTree<uint16_t>(xmpp_instance_count_,
                                     "DISCOVERY.max_control_nodes")) {
         xmpp_instance_count_ = 2;
@@ -588,7 +588,7 @@ void AgentParam::ParseVirtualHostArguments
 void AgentParam::ParseDiscoveryArguments
     (const boost::program_options::variables_map &var_map) {
     GetOptValue<string>(var_map, dss_server_, "DISCOVERY.server");
-    GetOptValue<uint16_t>(var_map, dss_port_, "DISCOVERY.port");
+    GetOptValue<uint32_t>(var_map, dss_port_, "DISCOVERY.port");
     GetOptValue<uint16_t>(var_map, xmpp_instance_count_,
                           "DISCOVERY.max_control_nodes");
 }
@@ -1199,7 +1199,7 @@ AgentParam::AgentParam(Agent *agent, bool enable_flow_options,
          "Run agent in vrouter / tsn / tor mode")
         ("DEFAULT.agent_base_directory", opt::value<string>(),
          "Base directory used by the agent")
-        ("DISCOVERY.port", opt::value<uint16_t>()->default_value(DISCOVERY_SERVER_PORT),
+        ("DISCOVERY.port", opt::value<uint32_t>()->default_value(DISCOVERY_SERVER_PORT),
          "Listen port of discovery server")
         ("DISCOVERY.server", opt::value<string>()->default_value("127.0.0.1"),
          "IP address of discovery server")
